Rename test package name and reable vpc test
