+++ pkg/apiserver/registry/inventory/vpc_suite_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package inventory_test
+package inventory
 
 import (
 	"testing"
